Step9/requirements.txt
pandas==1.4.0
pytest==6.2.5
requests==2.26.0
beautifulsoup4==4.10.0
mysql-connector-python==8.0.29
sqlalchemy==1.4.9
azure-storage==0.32.0
azure-storage-blob==12.12.0
